move some more css into separate classes
