+++ assets/pages/new-contributions-landing/components/ButtonWithRightArrow/ButtonWithRightArrow.scss
@@ -1,0 +1,12 @@
+// -----  gu-sass ----- //
+
+.button--right-arrow {
+  width: 100%;
+  padding: 0 50px 0 $gu-h-spacing;
+  text-align: left;
+
+  .svg-arrow-right-straight {
+    left: calc(100% - 34px);
+    height: 16px;
+  }
+}
