+++ assets/pages/new-contributions-landing/components/ContributionForm.jsx
@@ -47,7 +47,7 @@ import {
   setStripeV3HasLoaded,
 } from '../contributionsLandingActions';
 import ContributionErrorMessage from './ContributionErrorMessage';
-import StripePaymentRequestButtonContainer from './StripePaymentRequestButtonContainer';
+import StripePaymentRequestButtonContainer from './StripePaymentRequestButton/StripePaymentRequestButtonContainer';
 
 
 // ----- Types ----- //
