+++ assets/pages/new-contributions-landing/components/ContributionThankYou.jsx
@@ -7,9 +7,9 @@ import type { PaymentMethod } from 'helpers/contributions';
 import React from 'react';
 import { connect } from 'react-redux';
 import { type ContributionType, getSpokenType } from 'helpers/contributions';
-import { ButtonWithRightArrow } from './ButtonWithRightArrow';
+import { ButtonWithRightArrow } from './ButtonWithRightArrow/ButtonWithRightArrow';
 import MarketingConsent from './MarketingConsentContainer';
-import ContributionsSurvey from '../components/ContributionsSurvey';
+import ContributionsSurvey from './ContributionSurvey/ContributionsSurvey';
 import { type Action, setHasSeenDirectDebitThankYouCopy } from '../contributionsLandingActions';
 
 // ----- Types ----- //
