+++ assets/pages/new-contributions-landing/contributionsLanding.scss
@@ -11,6 +11,10 @@
 @import '~components/generalErrorMessage/generalErrorMessage';
 @import '~pages/new-contributions-landing/components/ContributionThankYouPasswordSet/ContributionThankYouPasswordSet';
 @import '~pages/new-contributions-landing/components/ContributionTicker/ContributionTicker';
+@import '~pages/new-contributions-landing/components/ButtonWithRightArrow/ButtonWithRightArrow';
+@import '~pages/new-contributions-landing/components/StripePaymentRequestButton/StripePaymentRequestButton';
+@import '~pages/new-contributions-landing/components/ContributionSurvey/ContributionSurvey';
+
 
 body {
   min-height: 100vh;
@@ -759,17 +763,6 @@ form {
   background-color: transparent;
 }
 
-.button--right-arrow {
-  width: 100%;
-  padding: 0 50px 0 $gu-h-spacing;
-  text-align: left;
-
-  .svg-arrow-right-straight {
-    left: calc(100% - 34px);
-    height: 16px;
-  }
-}
-
 .svg-subscribe, .svg-lock {
   width: 23px;
   height: 20px;
@@ -810,69 +803,6 @@ form {
   margin-top: $gu-v-spacing;
 }
 
-.component-contributions-survey {
-  border-bottom: 1px solid gu-colour(garnett-neutral-4);
-  margin: 0 -10px;
-  padding: 0 10px;
-
-  h3 {
-    font-family: $gu-text-egyptian-web;
-    letter-spacing: -0.5px;
-    font-weight: 900;
-    font-size: 17px;
-    margin: ($gu-v-spacing / 3) 0;
-  }
-
-  .component-contributions-survey__button {
-    display: inline-block;
-    margin-bottom: $gu-v-spacing * 3;
-    margin-top: $gu-v-spacing;
-
-    .button--survey {
-      background-color: gu-colour(garnett-neutral-1);
-      color: white;
-    }
-  }
-
-  .component-cta-link__text {
-    padding-right: 25px;
-    font-size: 16px;
-  }
-}
-
-
-/* Stripe Payment Request Button */
-
-.stripe-payment-request-button__button {
-  margin-top: $gu-v-spacing * 2;
-  margin-bottom: $gu-h-spacing;
-}
-
-.stripe-payment-request-button__divider {
-  line-height: 32px;
-  padding: 0;
-  position: relative;
-  overflow: hidden;
-  text-align: center;
-}
-.stripe-payment-request-button__divider:before,.stripe-payment-request-button__divider:after {
-  content:" ";
-  position: absolute;
-  top: 50%;
-  margin-left: -999em;
-  height: 2px;
-  width: 998em;
-  border-top: solid 1px gu-colour(garnett-neutral-4);
-}
-.stripe-payment-request-button__divider:after {
-  left: auto;
-  width: 999em;
-  margin: 0 0 0 1em;
-}
-
-.stripe-payment-request-button--hidden {
-  display: none;
-}
 
 /* Small header test */
 @include mq($from: mobile, $until: tablet) {
@@ -895,4 +825,3 @@ form {
     }
   }
 }
-
